CHIP OSP with a modified tableau method
I couldn't get the OSPs method of Heraldry to work at all, so I just coppied the normal method used by the base game, and switch the meshes to the one from the mod. the current problem with this is some of the Heraldry appears on the persons side. But does actually function correctly, the location of the Heraldry just needs to be changed for some of the items.